Refactored base Navbar component into snippets for easier readibility/maintainability
